requirements.txt
aiomysql==0.3.2
aiosmtplib==5.0.0
alembic==1.17.2
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asyncmy==0.2.10
bcrypt==5.0.0
blinker==1.9.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.121.1
fastapi-mail==1.6.1
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
orjson==3.11.4
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
PyMySQL==1.1.2
python-dotenv==1.2.1
python-jose==3.5.0
PyYAML==6.0.3
regex==2025.11.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
